+++ packages/react/src/components/F0Card/components/CardActions.tsx
@@ -3,6 +3,7 @@ import { Link, type LinkProps } from "@/components/Actions/Link"
 import { IconType } from "@/components/Utilities/Icon"
 import { cn } from "@/lib/utils"
 import { CardFooter } from "@/ui/Card"
+import { useMediaQuery } from "usehooks-ts"
 
 export interface CardPrimaryAction {
   label: string
@@ -32,6 +33,7 @@ export function CardActions({
   secondaryActions,
   compact = false,
 }: CardActionsProps) {
+  const isDesktop = useMediaQuery("(min-width: 640px)")
   const hasActions = primaryAction || secondaryActions
 
   if (!hasActions) {
@@ -41,24 +43,24 @@ export function CardActions({
   return (
     <CardFooter
       className={cn(
-        "justify-between gap-2 [&>div]:z-[1]",
+        "flex-col gap-2 sm:flex-row sm:justify-between [&>div]:z-[1]",
         "relative -mx-4 mt-4 border-0 border-t border-solid border-t-f1-border-secondary px-4 pt-4",
         compact && "pt-3"
       )}
     >
       {secondaryActions && (
-        <div className="flex gap-2">
+        <div className="flex w-full flex-col gap-2 sm:flex-row [&_a]:justify-center sm:[&_a]:justify-start [&_button]:w-full sm:[&_button]:w-fit [&_div]:w-full [&_div]:justify-center sm:[&_div]:w-fit">
           {Array.isArray(secondaryActions) ? (
             secondaryActions.map((action, index) => (
               <Button
                 key={index}
                 label={action.label}
                 icon={action.icon}
-                hideLabel={index > 0}
-                round={index > 0}
                 variant="outline"
                 onClick={action.onClick}
-                size={compact ? "sm" : "md"}
+                hideLabel={isDesktop && index > 0}
+                round={isDesktop && index > 0}
+                size={isDesktop ? (compact ? "sm" : "md") : "lg"}
               />
             ))
           ) : (
@@ -66,6 +68,7 @@ export function CardActions({
               href={secondaryActions.href}
               target={secondaryActions.target}
               disabled={secondaryActions.disabled}
+              data-testid="secondary-link"
             >
               {secondaryActions.label}
             </Link>
@@ -74,12 +77,15 @@ export function CardActions({
       )}
 
       {primaryAction && (
-        <Button
-          label={primaryAction.label}
-          icon={primaryAction.icon}
-          onClick={primaryAction.onClick}
-          size={compact ? "sm" : "md"}
-        />
+        <div className="w-full sm:w-fit [&_button]:w-full sm:[&_button]:w-fit [&_div]:w-full [&_div]:justify-center">
+          <Button
+            label={primaryAction.label}
+            icon={primaryAction.icon}
+            onClick={primaryAction.onClick}
+            size={isDesktop ? (compact ? "sm" : "md") : "lg"}
+            data-testid="primary-button"
+          />
+        </div>
       )}
     </CardFooter>
   )