+++ packages/react/src/components/exports.ts
@@ -1,5 +1,6 @@
 export * from "./Actions/exports"
 export * from "./Charts/exports"
+export * from "./F0Card"
 export * from "./layouts/exports"
 export * from "./OneFilterPicker/exports"
 export * from "./UpsellingKit/exports"